+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-## 6.7.6 (future release)
+## 6.7.6
 ### Refactor
 - Use SDK Bundle instead of the Kafka SDK
+### Chore
+- Remove nolint, add comments
 
 ## 6.7.5
 ## Refactor